requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.2
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=7.4.0
google-genai>=1.50.0
fpdf2>=2.7.0
pypdf>=5.0.0
